We get a Nullpointer Exception when rebalance is called
I have made a burstLeaf - how it works...

1.) Create a new node to house all values greater than magnitude and
populate it. 

2.) Shove magnitude-value into parent and re-sort because the value does
not always have to be the "largest" a smaller leaf (with larger
neighbours) could also bust.

3.) Tricky Dicky, after sorting the parent values we look for the value
again and get its position in order to point the reference to the new
leaf (could be error-prone if a middle leaf breaks.. needs checking)...

4.) Afterwards we re-check the criteria for the parent node... might
need to arrange this procedure along with the height method to have it
occur in a loop all the way to the root...food for thought.